[IMP] Upgrade to Odoo V14, make lookups work from any location (worldwide, removing FR limitations), remap fields and improve.
- Upgrade both coreff_base and coreff_creditsafe to Odoo v14.0 (remove @apione, @api.multi, view_type, always return value from computed fields etc)
- Update manifest authors
- Replace creditsafe mixins with res.partner model class
- Improve coreff_connector as below
- Lower pagesize from 200 to 10 max company records to prevent display crashing
- Handle spaces in company names by escaping correctly to prevent crashing
- Detect if country specified before searching company, if not then search all supported countries
- Correctly handle HeadOffice (officeType) parameter in more countries than just France
- Only store VAT and phone number if present to prevent crashes
- Add new field to res.partner named "creditsafe_activity_classification" to handle mainActivity (SIC code in UK)
- Remap creditsafe_activity_code and creditsafe_activity_description to the mainActivity in company_summary instead of principalActivity, as this only works in France. Now should work worldwide.
- Add different .gitignore to handle Visual Studio and also Python
- Convert "Last Judgement" [creditsafe_last_judgement_date] into an Odoo datetime field and parse the string to proper datetime. Plan to do with ALL datetime fields soon.
- Credit Rating [creditsafe_rating] was wrongly mapped to maxValue which is always 100 for the range you are in. Should be mapped to value which is the average.
- Add Contract Limit field [creditsafe_contract_limit]
