+++ coreff_base/models/res_partner.py
@@ -44,7 +44,6 @@ class ResPartner(models.Model):
         # TODO
         return
 
-    @api.one
     def create_from(self):
         # TODO
         logging.debug("CREATE FROM CALL")
@@ -56,7 +55,6 @@ class ResPartner(models.Model):
         rec._check_company_code()
         return rec
 
-    @api.multi
     def write(self, values):
         res = super(ResPartner, self).write(values)
         if (
@@ -94,7 +92,6 @@ class ResPartner(models.Model):
         for partner in self:
             partner.display_name = names.get(partner.id)
 
-    @api.multi
     def name_get(self):
         res = []
         for rec in self:
