+++ coreff_creditsafe/README.rst
@@ -7,4 +7,4 @@ from CreditSafe and add them to res_partner model.
 Of course, you need a CreditSafe account and set its properties
 into "financial informations" configuration menu.
 
-**Credits:**: Article 714, Infibail, Credit Safe
+**Credits**: Article 714, Infibail, Credit Safe, Chris Mann (Open User Systems)
