+++ coreff_creditsafe/views/res_partner_views.xml
@@ -37,6 +37,8 @@
                             <field name="creditsafe_incorporation_date" />
                             <field name="creditsafe_activity_code" />
                             <field name="creditsafe_activity_description" />
+                            <!--CHRIS MANN: Add field for mainActivity classification-->
+                            <field name="creditsafe_activity_classification" />
                             <field name="creditsafe_country" />
                         </group>
                         <group col="1">
@@ -46,9 +48,10 @@
                                 <field name="creditsafe_rating_short" />
                                 <field name="creditsafe_rating_long" />
                                 <field name="creditsafe_credit_limit" />
+                                <field name="creditsafe_contract_limit" />
                             </group>
                             <group string="Judgements">
-                                <field name="creditsafe_last_judgement_date" />
+                                <field name="creditsafe_last_change_date" />
                                 <field name="creditsafe_last_ccj_date" />
                                 <field name="creditsafe_number_of_directors" />
                             </group>
